Core Update! Update Snes9x to 1.6.0 and add in a bunch of fixes. (#1495)
* Optimize Release Schemes

Optimize Release Schemes by turning off the debug modes and disabling the GL/MTL api validation.

* make glClear calls work in Debug only

This reduces the per frame GPU resources used and the frame time from around 8ms to around 2ms on AppleTV Gen3 (and older hardware) when running Release builds.

* disable nativeScale on tvOS

This doesn't appear to get us anything.

* Re-Enable DoubleBuffering rendering support for Genesis Core.

* Optimize Release Schemes

* Update PVSettingsModel.swift

Sync up to develop to remove conflicts

* Initial 1.60 Snes9x Update

* More updates for 1.60

* Sync up iOS build w tvOS for Snes9x

* Remove unneeded header from compilation

* Last changes before PR.

+++ Cores/snes9x/PVSNES/SNES/SNESCore/sha256.cpp
@@ -13,8 +13,9 @@
 *********************************************************************/
 
 /*************************** HEADER FILES ***************************/
+#include <stdint.h>
 #include <stdlib.h>
-#include <memory.h>
+#include <string.h>
 
 /****************************** MACROS ******************************/
 #define ROTLEFT(a,b) (((a) << (b)) | ((a) >> (32-(b))))
@@ -35,7 +36,7 @@ typedef unsigned int  WORD;             /* 32-bit word, change to "long" for 16-
 typedef struct {
 	BYTE data[64];
 	WORD datalen;
-	unsigned long long bitlen;
+	uint64_t bitlen;
 	WORD state[8];
 } SHA256_CTX;
